app/requirements.txt
flask
flask-monitoringdashboard
gunicorn
h5py
keras
mlflow
numpy
pillow
pytest
torch
werkzeug

